Bugfix: new CSS `<style>` removing (background-)?color failed
